+++ ingot-service/ingot-pms/ingot-pms-provider/src/main/java/com/ingot/cloud/pms/service/domain/SysUserService.java
@@ -53,7 +53,7 @@ public interface SysUserService extends BaseService<SysUser> {
      *
      * @param id 用户ID
      */
-    void removeUserById(long id);
+    void removeUserById(int id);
 
     /**
      * 更新用户信息
